Added Fedora Instructions
So this is what I needed to do in order to get my fedora server up and running as a bridge, and I've added these instructions as if these problems only affect Fedora users, but I would be surprised if that's the case, because the alternative ports that are encouraged in this guide for ORPort and ServerTransportListenAddr are exactly what selinux is supposed to stop. Because of that, I'd imagine that this would function the same on CentOS SUSE and RHEL.

 As for why the fedora instructions install the executable into normal bin, that's because you run into selinux problems there too when it's in /usr/local/bin , but problems that were more complicated than my tiny selinux knowledge can deal with.

policycoreutils-python-utils is needed for semanage.