Use system package obfs4
You don't have to build obfs4 from source on Fedora since it is already packaged and installable with dnf.

However I'm not sure if you need run
```
/sbin/restorecon -v `which obfs4proxy`
```
or not. I did not run it and the port seems to be reachable (tested with https://bridges.torproject.org/scan/).

The package also includes a file `/etc/tor/obfs4.torrc` which maybe could be used to simplify the setup process even further.